requirements.txt
appnope==0.1.0
beautifulsoup4==4.6.0
boto3==1.4.4
botocore==1.5.78
certifi==2017.4.17
chardet==3.0.4
click==6.7
DateTime==4.2
decorator==4.0.11
docutils==0.13.1
gevent==1.2.2
greenlet==0.4.12
httplib2==0.10.3
idna==2.5
ipython==6.1.0
ipython-genutils==0.2.0
jedi==0.10.2
jmespath==0.9.3
mysql-connector-python-rf==2.2.2
oauthlib==2.0.2
olefile==0.44
packaging==16.8
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.2.0
prompt-toolkit==1.0.14
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
redis==2.10.5
requests==2.18.1
requests-oauthlib==0.8.0
s3transfer==0.1.10
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
urllib3==1.21.1
wcwidth==0.1.7
zope.interface==4.4.2
